Support for enabling and disabling modules, close Project-ARTist/meta#1
Modules are enabled by default and can be disabled if preconditions are not satisfied.
We use this when a module requires a codelib but creating the codelib environment fails (probably b/c no codelib was merged beforehand).
In this case, the module is simply disabled so the compiler does not terminate early and other modules can still be executed.